requirements.txt
numpy==1.24.1
pandas
matplotlib
scipy
scikit-learn
torcheval
torchmetrics
tqdm
seaborn
requests
scienceplots  
geopy